+++ src/main/java/mods/belgabor/amtweaker/mods/amt/loggers/CommandLogger.java
@@ -25,14 +25,19 @@
 package mods.belgabor.amtweaker.mods.amt.loggers;
 
 import minetweaker.MineTweakerAPI;
+import minetweaker.api.item.IItemStack;
 import minetweaker.api.minecraft.MineTweakerMC;
 import minetweaker.api.player.IPlayer;
 import minetweaker.api.server.ICommandFunction;
 import mods.defeatedcrow.api.charge.ChargeItemManager;
 import mods.defeatedcrow.api.charge.IChargeItem;
-import mods.defeatedcrow.api.recipe.IChargeIce;
-import mods.defeatedcrow.api.recipe.RecipeRegisterManager;
+import mods.defeatedcrow.api.recipe.*;
+import mods.defeatedcrow.recipe.BrewingRecipe;
+import net.minecraft.block.Block;
+import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
+import net.minecraftforge.fluids.Fluid;
+import net.minecraftforge.fluids.FluidStack;
 
 import java.util.Map;
 
@@ -48,6 +53,8 @@ public class CommandLogger implements ICommandFunction {
                 "    list chocolate recipes",
                 "/minetweaker amt heat",
                 "    list clay pan and iron plate heat sources",
+                "/minetweaker amt recipes",
+                "    list all AMT specific recipes to the minetweaker log",
                 "/minetweaker amt slag",
                 "    list slag loot"
         }, new CommandLogger());
@@ -55,11 +62,159 @@ public class CommandLogger implements ICommandFunction {
 
     private void logBoth(IPlayer player, String s) {
         MineTweakerAPI.logCommand(s);
-        player.sendChat(s);
+        if (player != null)
+            player.sendChat(s);
+    }
+
+    private void logBarrelRecipes(IPlayer player) {
+        for (Map.Entry<Fluid, Fluid> entry: BrewingRecipe.instance.recipeMap().entrySet()) {
+            if (player != null)
+                player.sendChat(getItemDeclaration(entry.getKey()) + " ---> " + getItemDeclaration(entry.getValue()));
+            MineTweakerAPI.logCommand("mods.amt.Pan.addChocolateRecipe(" + getItemDeclaration(entry.getValue()) + ", " + getItemDeclaration(entry.getKey()) + ");");
+        }
+    }
+
+    private void logChocolateRecipes(IPlayer player) {
+        for (Map.Entry<Object, ItemStack> entry: RecipeRegisterManager.chocoRecipe.getRecipeList().entrySet()) {
+            String input = getObjectDeclaration(entry.getKey());
+            if (player != null)
+                player.sendChat(input + " ---> " + getItemDeclaration(entry.getValue()));
+            MineTweakerAPI.logCommand("mods.amt.Pan.addChocolateRecipe(" + getItemDeclaration(entry.getValue()) + ", " + input + ");");
+        }
+    }
+
+    private void logEvaporatorRecipes() {
+        for (IEvaporatorRecipe recipe: RecipeRegisterManager.evaporatorRecipe.getRecipeList()) {
+            ItemStack input = recipe.getInput();
+            ItemStack output = recipe.getOutput();
+            FluidStack secondary = recipe.getSecondary();
+            String o = "mods.amt.Evaporator.addRecipe(";
+            if (output != null)
+                o += getItemDeclaration(output) + ", ";
+            if (secondary != null)
+                o += getItemDeclaration(secondary) + ", ";
+            o += getItemDeclaration(input);
+            if (!recipe.returnContainer())
+                o += ", false";
+            o += ");";
+            MineTweakerAPI.logCommand(o);
+        }
+    }
+
+    private void logIceRecipes() {
+        for (IIceRecipe recipe: RecipeRegisterManager.iceRecipe.getRecipeList()) {
+            String output = "mods.amt.IceMaker.addRecipe(" + getItemDeclaration(recipe.getOutput()) + ", " + getItemDeclaration(recipe.getInput());
+            ItemStack container = recipe.getContainer();
+            if (container != null)
+                output += ", " + getItemDeclaration(container);
+            output += ");";
+            MineTweakerAPI.logCommand(output);
+        }
+    }
+
+    private void logPanRecipes() {
+        for (IPanRecipe recipe: RecipeRegisterManager.panRecipe.getRecipeList()) {
+            ItemStack jpoutput = recipe.getOutputJP();
+            String o = "mods.amt.Pan.addRecipe(" + getItemDeclaration(recipe.getOutput());
+            if (jpoutput != null) {
+                o += ", " + getItemDeclaration(jpoutput);
+            }
+            o += ", " + getItemDeclaration(recipe.getInput()) + ", \"" + recipe.getTex() + "\", \"" + recipe.getDisplayName() + "\");";
+            MineTweakerAPI.logCommand(o);
+        }
+    }
+
+    private void logPlateRecipes() {
+        for (IPlateRecipe recipe: RecipeRegisterManager.plateRecipe.getRecipeList()) {
+            String o = "mods.amt.Plate.addRecipe(" + getItemDeclaration(recipe.getOutput()) + ", " + getItemDeclaration(recipe.getInput()) + ", " + recipe.cookingTime() + ", ";
+            if (recipe.useOvenRecipe()) {
+                o += "true";
+            } else {
+                o += "false";
+            }
+            MineTweakerAPI.logCommand(o + ");");
+        }
+    }
+
+    private void logProcessorRecipes() {
+        for (IProcessorRecipe recipe: RecipeRegisterManager.processorRecipe.getRecipes()) {
+            String o = "mods.amt.Processor.addRecipe(" + getItemDeclaration(recipe.getOutput());
+            float sec = recipe.getChance();
+            boolean ret = recipe.forceReturnContainer();
+            if ((recipe.getSecondary() != null) | (sec != 0) | ret) {
+                o += ", " + getItemDeclaration(recipe.getSecondary());
+            }
+            o += ", [";
+            boolean first = true;
+            for (Object x: recipe.getInput()) {
+                if (!first) {
+                    o += ", ";
+                } else {
+                    first = false;
+                }
+                o += getObjectDeclaration(x);
+            }
+            o += "], ";
+            if (recipe.isFoodRecipe()) {
+                o += "true";
+            } else {
+                o += "false";
+            }
+            if ((sec != 0) | ret) {
+                o += ", " + sec;
+            }
+            if (ret) {
+                o += ", true";
+            }
+            MineTweakerAPI.logCommand(o + ");");
+        }
+    }
+
+    private void logTeaRecipes() {
+        for (ITeaRecipe recipe: RecipeRegisterManager.teaRecipe.getRecipeList()) {
+            String o = "mods.amt.Processor.addRecipe(" + getItemDeclaration(recipe.getOutput());
+            ItemStack milk = recipe.getOutputMilk();
+            if (milk != null) {
+                o += ", " + getItemDeclaration(milk);
+            }
+            o += ", " + getItemDeclaration(recipe.getInput()) + ", \"" + recipe.getTex() + "\"";
+            if ((milk != null) && (recipe.getTex() != recipe.getMilkTex())) {
+                o += ", \"" + recipe.getMilkTex() + "\"";
+            }
+            MineTweakerAPI.logCommand(o + ");");
+        }
     }
 
     private String getItemDeclaration(ItemStack stack) {
+        if (stack == null) {
+            return "null";
+        }
         return MineTweakerMC.getIItemStack(stack).toString();
+    }
+    private String getItemDeclaration(Fluid stack) {
+        return "<liquid:" + stack.getName() + ">";
+    }
+    private String getItemDeclaration(FluidStack stack) {
+        return "<liquid:" + stack.getFluid().getName() + "> * " + stack.amount;
+    }
+    private String getObjectDeclaration(Object stack) {
+        if (stack instanceof String) {
+            return "<ore:" + (String) stack + ">";
+        } else if (stack instanceof ItemStack) {
+            return getItemDeclaration((ItemStack) stack);
+        } else if (stack instanceof Item) {
+            return getItemDeclaration(new ItemStack((Item) stack, 1));
+        } else if (stack instanceof Block) {
+            return getItemDeclaration(new ItemStack((Block) stack, 1));
+        } else if (stack instanceof Fluid) {
+            return getItemDeclaration((Fluid) stack);
+        } else if (stack instanceof FluidStack) {
+            return getItemDeclaration((FluidStack) stack);
+        } else if (stack == null) {
+            return "null";
+        } else {
+            return "?????";
+        }
     }
 
     @Override
@@ -81,6 +236,8 @@ public class CommandLogger implements ICommandFunction {
                 }
             } else if (arguments[0].equalsIgnoreCase("choco")) {
                 logBoth(player, "Chocolate Recipes:");
+                logChocolateRecipes(player);
+                /*
                 for (Map.Entry<Object, ItemStack> entry: RecipeRegisterManager.chocoRecipe.getRecipeList().entrySet()) {
                     String input = "";
                     if (entry.getKey() instanceof String) {
@@ -91,6 +248,7 @@ public class CommandLogger implements ICommandFunction {
                     player.sendChat(input + " ---> " + getItemDeclaration(entry.getValue()));
                     MineTweakerAPI.logCommand("mods.amt.Pan.addChocolateRecipe(" + getItemDeclaration(entry.getValue()) + ", " + input + ");");
                 }
+                */
             } else if (arguments[0].equalsIgnoreCase("heat")) {
                 logBoth(player, "Pan:");
                 for (ItemStack i: RecipeRegisterManager.panRecipe.getHeatSourceList()) {
@@ -100,6 +258,24 @@ public class CommandLogger implements ICommandFunction {
                 for (ItemStack i: RecipeRegisterManager.plateRecipe.getHeatSourceList()) {
                     logBoth(player, getItemDeclaration(i) + " -- " + i.getDisplayName());
                 }
+            } else if (arguments[0].equalsIgnoreCase("recipes")) {
+                MineTweakerAPI.logCommand("Barrel brewing recipes:");
+                logBarrelRecipes(null);
+                MineTweakerAPI.logCommand("Chocolate fondue recipes:");
+                logChocolateRecipes(null);
+                MineTweakerAPI.logCommand("Evaporator recipes:");
+                logEvaporatorRecipes();
+                MineTweakerAPI.logCommand("Ice Maker recipes:");
+                logIceRecipes();
+                MineTweakerAPI.logCommand("Pan recipes:");
+                logPanRecipes();
+                MineTweakerAPI.logCommand("Plate recipes:");
+                logPlateRecipes();
+                MineTweakerAPI.logCommand("Processor recipes:");
+                logProcessorRecipes();
+                MineTweakerAPI.logCommand("Teamaker recipes:");
+                logTeaRecipes();
+                player.sendChat("AMT recipes have been written to the MineTweaker log.");
             } else if (arguments[0].equalsIgnoreCase("slag")) {
                 for (int l=1; l <= 5; l++) {
                     logBoth(player, "Tier " + l + ":");